[APDV-111] Moved AddFieldRequestBodyMapper
