+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/field/SaveNewFieldImpl.java
@@ -2,7 +2,7 @@ package pl.edu.agh.apdvbackend.use_cases.field;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.edu.agh.apdvbackend.mappers.field.add_request_body.AddFieldRequestBodyMapper;
+import pl.edu.agh.apdvbackend.mappers.field.AddFieldRequestBodyMapper;
 import pl.edu.agh.apdvbackend.models.body_models.field.AddFieldRequestBody;
 import pl.edu.agh.apdvbackend.models.database.Field;
 import pl.edu.agh.apdvbackend.repositories.FieldRepository;
