+++ backend/src/test/java/pl/edu/agh/apdvbackend/mappers/field/AddFieldRequestBodyMapperTest.java
@@ -1,4 +1,4 @@
-package pl.edu.agh.apdvbackend.mappers.field.add_request_body;
+package pl.edu.agh.apdvbackend.mappers.field;
 
 import java.util.ArrayList;
 import java.util.Optional;
@@ -10,6 +10,7 @@ import org.springframework.boot.test.mock.mockito.SpyBean;
 import pl.edu.agh.apdvbackend.fakes.FieldFakes;
 import pl.edu.agh.apdvbackend.fakes.UnitFakes;
 import pl.edu.agh.apdvbackend.fakes.body_models.field.AddFieldRequestBodyFakes;
+import pl.edu.agh.apdvbackend.mappers.field.AddFieldRequestBodyMapper;
 import pl.edu.agh.apdvbackend.mappers.unit.UnitNameMapper;
 import pl.edu.agh.apdvbackend.models.database.FieldType;
